backend/requirements.txt
# Strategic Survey Engine - Backend Dependencies
# Python 3.11+

# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Config & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Auth (Phase 2+)
PyJWT>=2.8.0

# Export (Phase 6)
openpyxl>=3.1.0
reportlab>=4.0.0
